Split the EE embedding plugin registration in two
metabase-enterprise/embedding/index.ts registered the SDK data pickers at
import time and, in initializePlugin, both the embedding entitlement
predicates and the InteractiveEmbeddingSettingsCard admin component. Since
sdk-plugins.ts imported it, the SDK bundle's boot path reached the admin
settings card.

embedding-plugin.ts now registers the pickers and the entitlement
predicates, admin-settings-plugin.ts registers the settings card, and
has-embedding-feature.ts holds the premium-feature check both use. The
picker registration moved from module scope into initializePlugin, so it
runs from the composition roots at boot.
