+++ enterprise/frontend/src/metabase-enterprise/embedding/has-embedding-feature.ts
@@ -1,0 +1,3 @@
+import { hasPremiumFeature } from "metabase-enterprise/settings";
+
+export const hasEmbeddingFeature = () => hasPremiumFeature("embedding");
